+++ src/Samples/index.ts
@@ -4349,7 +4349,6 @@ let urlS2S;
         await rainbowSDK.stop();
     }
 
-
     rainbowSDK.start(token).then(async (result: any) => {
 //Promise.resolve({}).then(async(result: any) => {
         try {
@@ -4359,7 +4358,9 @@ let urlS2S;
             logger.log("debug", "MAIN - rainbow SDK started with result 1 : ", result); //logger.colors.green(JSON.stringify(result)));
             logger.log("debug", "MAIN - rainbow SDK started with credentials result 1 : ", logger.colors.green(connectedUser)); //logger.colors.green(JSON.stringify(result)));
 
+            let companyInfo = await rainbowSDK.contacts.getCompanyInfos();
 
+            logger.log("debug", "MAIN - company infos :" , companyInfo);
             /*
                     await rainbowSDK.stop().then((result)=>{
                         logger.log("debug", "MAIN - rainbow SDK stop : ", result); //logger.colors.green(JSON.stringify(result)));